added createPNG & writePNG

+++ src/crc.cc
@@ -1,5 +1,7 @@
+#include "crc.hpp"
+
 /* Table of CRCs of all 8-bit messages. */
-static unsigned long crc_table[256];
+static uint32_t crc_table[256];
 
 /* Flag: has the table been computed? Initially false. */
 static bool crc_table_computed = false;
@@ -7,11 +9,11 @@ static bool crc_table_computed = false;
 /* Make the table for a fast CRC. */
 static void make_crc_table(void)
 {
-  unsigned long c;
+  uint32_t c;
   int n, k;
 
   for (n = 0; n < 256; n++) {
-    c = (unsigned long) n;
+    c = (uint32_t) n;
     for (k = 0; k < 8; k++) {
       if (c & 1)
         c = 0xedb88320L ^ (c >> 1);
@@ -28,10 +30,10 @@ static void make_crc_table(void)
    is the 1's complement of the final running CRC (see the
    crc() routine below). */
 
-static unsigned long update_crc(unsigned long crc, unsigned char *buf,
+static uint32_t update_crc(uint32_t crc, unsigned char *buf,
                          int len)
 {
-  unsigned long c = crc;
+  uint32_t c = crc;
   int n;
 
   if (!crc_table_computed)
@@ -43,7 +45,7 @@ static unsigned long update_crc(unsigned long crc, unsigned char *buf,
 }
 
 /* Return the CRC of the bytes buf[0..len-1]. */
-unsigned long crc(unsigned char *buf, int len)
+uint32_t crc(unsigned char *buf, int len)
 {
   return update_crc(0xffffffffL, buf, len) ^ 0xffffffffL;
 }
